+++ library/Morloc/Database/Typecheck.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE OverloadedStrings #-}
+
 {-|
 Module      : Morloc.Database.Typecheck
 Description : Check the logical consistency of a program
@@ -5,11 +7,17 @@ Copyright   : (c) Zebulun Arendsee, 2018
 License     : GPL-3
 Maintainer  : zbwrnz@gmail.com
 Stability   : experimental
+
+This typechecker currently only checks the case where an argument with a
+concrete type receives input from a function with a concrete output. In this
+case, the typechecker raises a warning (but does not die) if the types are not
+identical.
 -}
 
 module Morloc.Database.Typecheck (typecheck) where
 
 import Morloc.Types
+import Morloc.Operators
 import qualified Morloc.Error as ME
 import qualified Morloc.Data.Text as MT
 import qualified Morloc.Component.Manifold as Manifold
@@ -34,12 +42,28 @@ compareTypes :: (MType, Argument) -> IO ()
 compareTypes (parentType, (ArgCall c)) = case mAbstractType c of
   (Just (MFuncType _ _ childOutput)) ->
     if
-      parentType == childOutput 
+      eqTypes parentType childOutput 
     then
-      return () 
-    else
       return ()
-      -- fail (show $ ME.TypeConflict (MT.unpack $ MT.pretty parentType)
-      --                              (MT.unpack $ MT.pretty childOutput))
+    else
+      do
+        MT.putStr (MT.show' $ ME.TypeConflict (showType parentType)
+                                              (showType childOutput))
+        return ()
   _ -> return ()
 compareTypes _ = return ()
+
+showType :: MType -> MT.Text
+showType (MConcType _ x xs) = MT.intercalate " " ([MT.pretty x] ++ map showType xs)
+showType (MAbstType _ x xs) = MT.intercalate " " ([MT.pretty x] ++ map showType xs)
+showType (MFuncType _ xs o) = MT.intercalate ", " (map showType xs) <> " -> " <> MT.pretty o
+
+-- TODO: This is VERY rudimentary. I am treating generics as completely
+-- unconstrainted wildcards. Instead, I should resolve all generic cases to
+-- concrete cases first, raising errors for any ambiguities.
+eqTypes :: MType -> MType -> Bool
+eqTypes (MAbstType _ _ _) _ = True
+eqTypes _ (MAbstType _ _ _) = True
+eqTypes (MConcType _ x xs) (MConcType _ y ys)
+  =  x == y
+     && foldl (&&) True (zipWith eqTypes xs ys)