+++ library/Morloc/Error.hs
@@ -52,8 +52,8 @@ errmsg (InvalidRDF msg)      = "Invalid RDF: " <> MT.show' msg
 errmsg (NotImplemented msg)  = "Not yet implemented: " <> MT.show' msg
 errmsg (NotSupported msg)    = "NotSupported: " <> MT.show' msg
 errmsg (SyntaxError err)     = "SyntaxError: " <> MT.show' err
-errmsg (TypeConflict t1 t2)  = MT.unlines 
-  [ "TypeConflict: expected type", t1, "but got", t2 ] 
+errmsg (TypeConflict t1 t2)  = 
+  "TypeConflict: cannot cast " <> t1 <> " as " <> t2 <> "\n"
 
 error' :: MT.Text -> a
 error' x = error $ MT.unpack x